Split pie trace into multiple files
