+++ src/traces/pie/layout_attributes.js
@@ -1,0 +1,18 @@
+/**
+* Copyright 2012-2016, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+'use strict';
+
+module.exports = {
+    /**
+     * hiddenlabels is the pie chart analog of visible:'legendonly'
+     * but it can contain many labels, and can hide slices
+     * from several pies simultaneously
+     */
+    hiddenlabels: {valType: 'data_array'}
+};
